+++ docs/about/whats_new.rst
@@ -8,6 +8,8 @@ What's New
 v1.8.next
 =========
 
+- Add `grid_spec` to `list_products` (:pull:`1357`)
+
 v1.8.9 (17 November 2022)
 =========================